cancel today's decal due to bad air quality
Chancellor Christ, I would like to express my utmost gratitude and
admiration for your efforts to continue classes during this
unprecedented period of statewide conflagration. UC Berkeley is a
fine institution, but has been plagued with an increasing student
population for decades. As such, I commend your wholehearted effort
to reduce campus crowding by subjecting students to potentially life
threatening air quality on a regular basis. The myriad of problems
facing Berkeley residents today would take several years of dedicated
planning and policy intervention to deal with, especially as the
population of the campus and surrounding community continues to grow
at such a drastic rate. However, by purging the weak-willed and the
weak-lunged, the same result can be achieved simply by reducing the
student body to a far more manageable level. The sacrifices of the
fallen will not be in vein—as each death will lead us to a better
Berkeley...and a better tomorrow. The souls of the discarded students
will naturally be offered to our campus deity, Oski, and his resulting
appeasement will bring forth a bountiful harvest of memes and boba.
Long live the ash! Death to the breathers!
FAITH IN CHRIST! UNIVERSITAS VULT!